# Cat charity/openapi_part_2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Благотворительный фонд поддержки котиков QRKot",
    "description": "Сервис для поддержки котиков",
    "version": "0.1.0"
  },
  "paths": {
...
    "/donation/": {
      "get": {
        "tags": [
          "donations"
        ],
        "summary": "Get All Donations",
        "description": "Показать список всех пожертвований.\n\nТолько для суперюзеров.",
        "operationId": "get_all_donations_donation__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DonationFullInfoDB"
                  },
                  "type": "array",
                  "title": "Response Get All Donations Donation  Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "403": {
            "description": "Not a superuser.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "donations"
        ],
        "summary": "Create Donation",
        "description": "Сделать пожертвование.\n\nТолько для зарегистрированных пользователей.",
        "operationId": "create_donation_donation__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DonationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DonationDB"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/donation/my": {
      "get": {
        "tags": [
          "donations"
        ],
        "summary": "Get User Donations",
        "description": "Показать список пожертвований пользователя, выполняющего запрос.\n\nТолько для зарегистрированных пользователей.",
        "operationId": "get_user_donations_donation_my_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DonationDB"
                  },
                  "type": "array",
                  "title": "Response Get User Donations Donation My Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DonationCreate": {
        "properties": {
          "full_amount": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Full Amount"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "full_amount"
        ],
        "title": "DonationCreate"
      },
      "DonationDB": {
        "properties": {
          "full_amount": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Full Amount"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "create_date": {
            "type": "string",
            "format": "date-time",
            "title": "Create Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "full_amount",
          "id",
          "create_date"
        ],
        "title": "DonationDB"
      },
      "DonationFullInfoDB": {
        "properties": {
          "full_amount": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Full Amount"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "create_date": {
            "type": "string",
            "format": "date-time",
            "title": "Create Date"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "invested_amount": {
            "type": "integer",
            "title": "Invested Amount"
          },
          "fully_invested": {
            "type": "boolean",
            "title": "Fully Invested"
          },
          "close_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "full_amount",
          "id",
          "create_date",
          "user_id",
          "invested_amount",
          "fully_invested"
        ],
        "title": "DonationFullInfoDB"
      },
...
      "ForbiddenError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ForbiddenError"
      },
      "UnauthorizedError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "UnauthorizedError"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}